+++ internal/command/command_catalog.go
@@ -106,9 +106,9 @@ func CommandsCatalog() map[string]cli.CommandFactory {
 		"status service": func() (cli.Command, error) {
 			return &commands.Service{Status: true}, nil
 		},
-		// Verbs for `service-group` resource
-		"create service-group": func() (cli.Command, error) {
-			return &commands.ServiceGroup{Create: true}, nil
+		// Verbs for `profile` resource
+		"create profile": func() (cli.Command, error) {
+			return &commands.Profile{Create: true}, nil
 		},
 
 		/*
